# SentinelOps-AI-Command-Center/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  systemHealth: number;
  activeIncidents: number;
  servicesHealthy: number;
  servicesTotal: number;
  aiConfidence: number;
  errorRate: number;
  recoveryTime: number;
}

export interface Incident {
  id: string;
  title: string;
  severity: string;
  status: string;
  workflowState: string;
  affectedServices: string[];
  rootCause: string;
  recommendation: string;
  risk: string;
  confidence: number;
  reversible: boolean;
  rollbackAvailable: boolean;
  /** @nullable */
  approvedBy?: string | null;
  /** @nullable */
  outcome?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface IncidentActionInput {
  actor?: string;
  note?: string;
}

export interface LogEntry {
  timestamp: string;
  service: string;
  level: string;
  message: string;
}

export interface IncidentMetrics {
  cpu: number;
  memory: number;
  latency: number;
  errorRate: number;
  connections: number;
}

export interface TraceSpan {
  service: string;
  dependency: string;
  latency: number;
  status: string;
}

export interface TimelineEvent {
  timestamp: string;
  type: string;
  description: string;
  service: string;
  status: string;
}

export interface AuditEvent {
  id: string;
  timestamp: string;
  actor: string;
  action: string;
  target: string;
  result: string;
  risk: string;
  incidentId: string;
}

export interface EvaluationScenario {
  id: string;
  name: string;
  groundTruth: string;
  agentRootCause: string;
  baselineRootCause: string;
  agentConfidence: number;
  diagnosisSeconds: number;
  remediationSuccess: boolean;
}

export interface EvaluationMetrics {
  rootCauseAccuracy: number;
  timeToDiagnosis: number;
  remediationSuccessRate: number;
  falsePositiveRate: number;
  confidenceCalibration: number;
}

export interface EvaluationResults {
  scenarios: EvaluationScenario[];
  agent: EvaluationMetrics;
  baseline: EvaluationMetrics;
}