# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/auditEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEvent {
  id: string;
  timestamp: string;
  actor: string;
  action: string;
  target: string;
  result: string;
  risk: string;
  incidentId: string;
}