# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  systemHealth: number;
  activeIncidents: number;
  servicesHealthy: number;
  servicesTotal: number;
  aiConfidence: number;
  errorRate: number;
  recoveryTime: number;
}