# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/evaluationMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EvaluationMetrics {
  rootCauseAccuracy: number;
  timeToDiagnosis: number;
  remediationSuccessRate: number;
  falsePositiveRate: number;
  confidenceCalibration: number;
}