# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/evaluationResults.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationMetrics } from './evaluationMetrics';
import type { EvaluationScenario } from './evaluationScenario';

export interface EvaluationResults {
  scenarios: EvaluationScenario[];
  agent: EvaluationMetrics;
  baseline: EvaluationMetrics;
}