# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/evaluationScenario.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EvaluationScenario {
  id: string;
  name: string;
  groundTruth: string;
  agentRootCause: string;
  baselineRootCause: string;
  agentConfidence: number;
  diagnosisSeconds: number;
  remediationSuccess: boolean;
}