# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/incident.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Incident {
  id: string;
  title: string;
  severity: string;
  status: string;
  workflowState: string;
  affectedServices: string[];
  rootCause: string;
  recommendation: string;
  risk: string;
  confidence: number;
  reversible: boolean;
  rollbackAvailable: boolean;
  /** @nullable */
  approvedBy?: string | null;
  /** @nullable */
  outcome?: string | null;
  createdAt: string;
  updatedAt: string;
}