# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/incidentMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IncidentMetrics {
  cpu: number;
  memory: number;
  latency: number;
  errorRate: number;
  connections: number;
}