# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/logEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LogEntry {
  timestamp: string;
  service: string;
  level: string;
  message: string;
}