# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  timestamp: string;
  type: string;
  description: string;
  service: string;
  status: string;
}