# SentinelOps-AI-Command-Center/lib/api-zod/src/generated/types/traceSpan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TraceSpan {
  service: string;
  dependency: string;
  latency: number;
  status: string;
}